Small refactor of UserService and UserController On branch Add_List_all_Users Changes to be committed: 	modified:   src/main/java/keysat/controller/UserController.java 	renamed:    src/main/java/keysat/UserService.java -> src/main/java/keysat/service/UserService.java 	modified:   src/test/java/keysat/controller/UserControllerTest.java
